requirements.txt
# Core async
aiohttp>=3.9.0
websockets>=12.0

# Data
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0  # For parquet support

# Database
asyncpg>=0.29.0
redis>=5.0.0
influxdb-client>=1.38.0

# Monitoring (optional)
prometheus-client>=0.19.0

# Config
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Code quality (dev)
ruff>=0.1.0
mypy>=1.7.0
